requirements.txt
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==4.4.2
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
mutagen==1.46.0
numpy==1.24.3
Pillow==9.5.0
proglog==0.1.10
pycryptodomex==3.17
python-dotenv==1.0.0
pytube==15.0.0
requests==2.30.0
tqdm==4.65.0
urllib3==2.0.2
vtt-to-srt3==0.2.0.1
websockets==11.0.2
yt-dlp==2023.3.4
